// types/chat.ts
export enum PrivateRoom {
  SEND = "room-message",
  RECEIVE = "receive-room-message",
  ENTER = "enter-room",
  SUCCESS = "enter-room-success",
  EXIT = "exit-room",
  NOTICE = "leave-message",
}

export enum PublicRoom {
  MSG = "receive-message",
  SEND = "send-message",
}

export enum LostArk {
  CHA = "character",
}